fix state where nvm will never install
if $NVM_DIR is set by the bash_profile
good-morning adds, but the dir was not
created yet, nvm will not install.
This may be due to an early exit of the script, which is normal during
the first run where the system
may require some system updates and
a reboot.